+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Move ISONE Monthly Asset NCPC to DuckDB
 - test
 
+# Release 2026-03-18
+- Fix urls in the extension methods for LMPs
+
 # Release 2026-03-17
 - Add extension methods to calculate daily and monthly LMPs to several ISOs.